app/requirements.txt
requests==2.31
python-decouple==3.8
apprise==1.5.0